Refactoring, move extensions closer to their use-case.
